Beginning of splitting into components
